LayoutTests/sputnik/Conformance/12_Statement/12.14_The_try_Statement/S12.14_A16_T9-expected.txt

CONSOLE MESSAGE: line 77: SyntaxError: Parse error
S12.14_A16_T9

PASS Expected parsing failure

TEST COMPLETE
